testing + tests: Pass single value to range test functions
This makes range tests a bit nicer because you no longer have to write the loop
yourself in every range test. The loop is now in a helper function defined
inside the RANDOM_TEST macro. Putting the loop into run_range_test instead would
lead to a ~10% performance hit because of the indirect function call overhead.