charconv: Implement from_chars
This implements the from_chars functions similar to std::from_chars in C++.
Unlike the C++ versions these functions can optionally autodetect the base
according to the prefix (e.g. 0x prefix means base 16). This commit also
includes fortification and extensive tests, but no benchmark for now. The
implementation is not very optimized yet, but it is efficient enough and the API
is much more convenient than the C standard library functions (strtol and such).