charconv: Print plus/minus sign for all bases
There is no good reason to special case base 10. This brings us closer to
std::to_chars. We used to do this because you usually don't want a sign for
bases other than 10. But if you don't want a minus sign for a negative value
just cast it to an unsigned value/use one of the unsigned functions.
Also default to base 10 if base is 0 but there are other flags.