testing: Print output of failed tests for single tests
This copies the approach from testing.c for single_testing.c. We temporarily
redirect stdout and stderr to a memfd and then print the content of the memfd if
the test failed. It is not enough to make test_log print to a fmemopen (or
rather open_memstream) file like the TODO said, because fortify tests for
example print an error message directly to stderr.